renamed queue -> edge
