+++ limen-core/src/prelude.rs
@@ -1,19 +1,23 @@
 //! Convenience re-exports for implementers.
 
+pub use crate::edge;
 pub use crate::errors::*;
 pub use crate::graph::*;
 pub use crate::memory::*;
 pub use crate::message::*;
-pub use crate::node::*;
+pub use crate::node;
 pub use crate::platform::*;
 pub use crate::policy::*;
-pub use crate::queue::*;
 pub use crate::scheduling::*;
 pub use crate::telemetry::*;
 pub use crate::types::*;
 
 // Used by define_graph macro.
 // pub use crate::define_graph;
+pub use crate::edge::{
+    link::{EdgeDescriptor, EdgeLink},
+    NoQueue,
+};
 pub use crate::graph::{
     GraphApi, GraphEdgeAccess, GraphNodeAccess, GraphNodeContextBuilder, GraphNodeTypes,
 };
@@ -22,8 +26,4 @@ pub use crate::node::{
     StepContext,
 };
 pub use crate::policy::EdgePolicy;
-pub use crate::queue::{
-    link::{EdgeDescriptor, EdgeLink},
-    NoQueue,
-};
 pub use crate::types::{EdgeIndex, NodeIndex, PortId, PortIndex};
